backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-docx==1.1.0
python-multipart==0.0.6
anthropic>=0.40.0
python-dotenv==1.0.0

